- integrate some of the changes that were not ported over to this public repo, some were already implemented - fix some test nits

+++ Samples/VirtualCamera/VirtualCameraTest/EVRHelper.cpp
@@ -16,7 +16,7 @@ EVRHelper::~EVRHelper()
 {
     if (m_spEVRSinkWriter)
     {
-        m_spEVRSinkWriter->Flush(MF_SINK_WRITER_ALL_STREAMS);
+        m_spEVRSinkWriter->Flush((DWORD)MF_SINK_WRITER_ALL_STREAMS);
         m_spEVRSinkWriter = nullptr;
     }
 }
